// src/types/api/documentFields.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Pet Management API
 * Pet Management API with health logs, prescriptions, item master, and JWT authentication.
 * OpenAPI spec version: 2
 */
import type { DocumentFieldsBody } from "./documentFieldsBody";

export interface DocumentFields {
  parentDocId?: string;
  title: string;
  /** JSONB body for block-based structure */
  body?: DocumentFieldsBody;
}